# src/api/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Numbers": {
				"properties": {
					"OBITUARY": {
						"type": "number",
						"format": "double"
					},
					"POLITICS": {
						"type": "number",
						"format": "double"
					},
					"NOTICE": {
						"type": "number",
						"format": "double"
					},
					"PUBLIC": {
						"type": "number",
						"format": "double"
					},
					"OTHER": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Type Numbers contains the total number of board registered for each `Type`."
			},
			"Point": {
				"description": "Type Point represents a GeoJSON point.",
				"properties": {
					"type": {
						"type": "string",
						"enum": [
							"Point"
						],
						"nullable": false,
						"description": "GeoJSON type. Is always \"Point\".",
						"default": "Point"
					},
					"coordinates": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "GeoJSON coordinates. Longitude is the first element and latitude follows.",
						"example": [
							-103.4,
							-45.7
						],
						"default": [],
						"minItems": 2,
						"maxItems": 2
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Accessibility": {
				"description": "Enum Accessibility represents a board's physical environment and reachability.",
				"enum": [
					"PUBLIC",
					"PRIVATE",
					"INACCESSIBLE"
				],
				"type": "string"
			},
			"Material": {
				"description": "Enum Material describes a board's surface material.",
				"enum": [
					"WOOD",
					"METAL",
					"PLASTIC",
					"GLASS",
					"OTHER"
				],
				"type": "string"
			},
			"Size": {
				"description": "Enum Size describes a boards's general physical size.",
				"enum": [
					"SMALL",
					"MEDIUM",
					"LARGE"
				],
				"type": "string"
			},
			"Type": {
				"description": "Enum Type represents a board's purpose.",
				"enum": [
					"OBITUARY",
					"POLITICS",
					"NOTICE",
					"PUBLIC",
					"OTHER"
				],
				"type": "string"
			},
			"Board": {
				"description": "Type Board represents a bulletin board.",
				"properties": {
					"location": {
						"$ref": "#/components/schemas/Point",
						"description": "Board position as GeoJSON Point (see\n[geospatial queries](https://www.mongodb.com/docs/upcoming/geospatial-queries/)\nand [type definition](https://www.mongodb.com/docs/manual/reference/geojson/)).\nLongitude is the first element and latitude follows."
					},
					"accessibility": {
						"$ref": "#/components/schemas/Accessibility",
						"description": "Board accessibility."
					},
					"material": {
						"$ref": "#/components/schemas/Material",
						"description": "Board material."
					},
					"size": {
						"$ref": "#/components/schemas/Size",
						"description": "Board size."
					},
					"traffic": {
						"type": "integer",
						"format": "int32",
						"description": "Board traffic amount.",
						"minimum": 1,
						"maximum": 5
					},
					"type": {
						"$ref": "#/components/schemas/Type",
						"description": "Board type."
					},
					"created": {
						"type": "string",
						"format": "date-time",
						"description": "Board creation date."
					},
					"modified": {
						"type": "string",
						"format": "date-time",
						"description": "Board last modification date."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Filter": {
				"description": "Type Filter represents a transient object used to query the API for boards.",
				"properties": {
					"accessibility": {
						"items": {
							"$ref": "#/components/schemas/Accessibility"
						},
						"type": "array",
						"description": "Filter the results' accessibility, if defined.",
						"example": [
							"PUBLIC",
							"PRIVATE"
						],
						"maxItems": 3
					},
					"material": {
						"items": {
							"$ref": "#/components/schemas/Material"
						},
						"type": "array",
						"description": "Filter the results' material, if defined.",
						"example": [
							"GLASS",
							"WOOD",
							"OTHER"
						],
						"maxItems": 5
					},
					"size": {
						"items": {
							"$ref": "#/components/schemas/Size"
						},
						"type": "array",
						"description": "Filter the results' size, if defined.",
						"example": [
							"SMALL",
							"LARGE"
						],
						"maxItems": 3
					},
					"traffic": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "Filter the results' traffic amount, if defined.",
						"example": [
							2,
							5
						],
						"minimum": 1,
						"maximum": 5
					},
					"type": {
						"items": {
							"$ref": "#/components/schemas/Type"
						},
						"type": "array",
						"description": "Filter the results' type, if defined.",
						"example": [
							"OBITUARY",
							"PUBLIC"
						],
						"maxItems": 5
					},
					"box": {
						"items": {
							"items": {
								"type": "number",
								"format": "double"
							},
							"type": "array"
						},
						"type": "array",
						"description": "Longitude and latitude \"box\" for the results. All boards returned will be\nlocated inside this box. Each of the array items is the tuple\n`[longitude, latitude]`, akin to GeoJSON points.",
						"example": [
							[
								-103.4,
								-45.7
							],
							[
								95.8,
								23.2
							]
						],
						"minItems": 2,
						"maxItems": 2
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "bakeka-backend",
		"version": "0.0.1",
		"description": "The backend monolith for Bakeka",
		"license": {
			"name": "AGPL-3.0-or-later"
		},
		"contact": {
			"name": "Bakeka"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/boards/numbers": {
			"get": {
				"operationId": "GetNumbers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Numbers"
								}
							}
						}
					}
				},
				"description": "Returns the total number of board registered for each type (see enum Type).",
				"security": [],
				"parameters": []
			}
		},
		"/boards": {
			"post": {
				"operationId": "FilterBoards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Board"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves multiple board by using various filters. If any of the properties\nof the filter is `undefined`, it will not be applied to the results.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The filter to apply",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Filter",
								"description": "The filter to apply"
							}
						}
					}
				}
			}
		},
		"/boards/{boardId}": {
			"get": {
				"operationId": "GetBoard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Board"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Retrieves a single board by its internal ID.",
				"security": [],
				"parameters": [
					{
						"description": "The board's identifier",
						"in": "path",
						"name": "boardId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "UpdateBoard",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updated a single board by its internal ID.",
				"security": [],
				"parameters": [
					{
						"description": "The board's identifier",
						"in": "path",
						"name": "boardId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The new board object",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Board",
								"description": "The new board object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}